requirements.txt
aniso8601==9.0.1
asgiref==3.7.2
Django==4.0.10
django-filter==23.4
djangorestframework==3.14.0
graphene==3.3
graphene-django==3.1.5
graphql-core==3.2.3
graphql-relay==3.2.0
promise==2.3
psycopg2-binary==2.9.7
pycparser==2.21
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
